basic functionality comparing articles by source bias from allsides.com database

+++ Sample/src/ArticleComparisonIndexer.java
@@ -1,31 +1,55 @@
 /**
- * Class for comparing the perspectives of two articles based on source bias and article text
+ * Class for comparing the perspectives of two articles based on source bias and
+ * article text
+ * 
  * @author jd
  *
  */
 public class ArticleComparisonIndexer {
 	private Article article1, article2;
-	private double index = 0;
-	
+	private int index = 0;
+
+	/**
+	 * 
+	 * @param a1
+	 *            - first article
+	 * @param a2
+	 *            - second article
+	 */
 	public ArticleComparisonIndexer(Article a1, Article a2) {
 		article1 = a1;
 		article2 = a2;
 	}
-	
-	public double getIndex() {
-		int sourceIndex =  getSourceIndex();
+
+	/**
+	 * 
+	 * @return - index representing how likely two articles are to differ in
+	 *         perspectives
+	 */
+	public int getIndex() {
+		int sourceIndex = getSourceIndex();
 		double textIndex = getTextIndex();
 		index = sourceIndex;
 		// enter formula here modifying index variable
 		return index;
 	}
-	
+
+	/**
+	 * 
+	 * @return - calculates difference index of two articles based on source bias
+	 *         database from allsides.com
+	 */
 	public int getSourceIndex() {
-		return Math.abs(article1.getIntRating() -  article2.getIntRating());
+		return Math.abs(article1.getIntRating() - article2.getIntRating());
 	}
-	
-	public double getTextIndex() {
-		//some natural language processing here
+
+	/**
+	 * 
+	 * @return - calculates difference index from analyzing phrases in text of two
+	 *         articles
+	 */
+	public int getTextIndex() {
+		// some natural language processing here
 		return 0;
 	}
 }
